requirements.txt
anytree==2.13.0
GitPython==3.1.45
python-gitlab==7.0.0
globre==0.1.5
PyYAML==6.0.3
rich==14.2.0
typer==0.12.5
pydantic==2.9.2
pydantic-settings==2.7.1
urllib3==2.3.0
